backend/data-ingestion/requirements.txt
fastapi
uvicorn
kafka-python
pydantic
prometheus-fastapi-instrumentator